requirements.txt
streamlit 
langchain
google-generativeai
python-dotenv